// src/enums/LayoutActionTypes.ts
enum LayoutActionTypes {
  CloseAsteroids = '@layout/CLOSE_ASTEROIDS',
  CloseMenu = '@layout/CLOSE_MENU',
  CloseTerminal = '@layout/CLOSE_TERMINAL',
  OpenAsteroids = '@layout/OPEN_ASTEROIDS',
  OpenMenu = '@layout/OPEN_MENU',
  OpenTerminal = '@layout/OPEN_TERMINAL',
  SetBarrelRolling = '@layout/SET_BARREL_ROLLING',
  SetTitle = '@layout/SET_TITLE',
  SetMenuItem = '@layout/SET_MENU_ITEM',
}

export default LayoutActionTypes;